Refactor batch operations and remove dependency resolver
- Removed the dependency resolver and batch operations subgraph to streamline the codebase.
- Updated the commit operations to include approval gates and job polling.
- Adjusted the README and documentation to reflect the removal of batch operations and the updated commit workflow.
- Modified workflows to eliminate references to batch operations, ensuring clarity in the available features.
- Updated tests and fixtures to align with the new structure and removed obsolete tests related to dependency resolution.
