+++ python/ai-agent-panos/pyproject.toml
@@ -37,6 +37,9 @@ panos-agent = "src.cli.commands:app"
 requires = ["hatchling"]
 build-backend = "hatchling.build"
 
+[tool.hatch.build.targets.wheel]
+packages = ["src"]
+
 [tool.black]
 line-length = 100
 target-version = ["py311"]
