--- python/ai-agent-panos/src/core/subgraphs/batch.py
@@ -1,373 +1,0 @@
-"""Batch operations subgraph with parallel execution and dependency resolution.
-
-Workflow:
-1. Validate batch input
-2. Resolve dependencies → split into levels
-3. Process each level:
-   - Fan-out: Send each item to parallel CRUD execution
-   - Fan-in: Aggregate results
-4. Move to next level or finish
-
-Key features:
-- Parallel execution within dependency levels
-- Sequential execution across levels
-- Prevents "not a valid reference" errors
-- 4-10x speedup for bulk operations
-"""
-
-import logging
-import uuid
-from typing import Literal
-
-from langgraph.graph import StateGraph, END, Send
-from langgraph.constants import Send as SendType
-
-from src.core.state_schemas import BatchState
-from src.core.dependency_resolver import sort_items_by_dependencies, get_dependency_summary
-from src.core.subgraphs.crud import create_crud_subgraph
-
-logger = logging.getLogger(__name__)
-
-
-def validate_batch_input(state: BatchState) -> BatchState:
-    """Validate batch operation inputs.
-
-    Args:
-        state: Current batch state
-
-    Returns:
-        Updated state with validation result
-    """
-    logger.info(
-        f"Validating batch {state['operation_type']} "
-        f"for {len(state['items'])} {state['object_type']} objects"
-    )
-
-    # Check required fields
-    if not state.get("items"):
-        return {
-            **state,
-            "result_message": "❌ Error: No items provided for batch operation",
-        }
-
-    if state["operation_type"] not in ["create", "update", "delete"]:
-        return {
-            **state,
-            "result_message": f"❌ Error: Unsupported batch operation: {state['operation_type']}",
-        }
-
-    # Validate each item has required fields
-    for i, item in enumerate(state["items"]):
-        if not item.get("name"):
-            return {
-                **state,
-                "result_message": f"❌ Error: Item {i} missing 'name' field",
-            }
-
-    return {
-        **state,
-        "total_items": len(state["items"]),
-        "completed_items": 0,
-        "successful_items": 0,
-        "failed_items": 0,
-        "failure_details": [],
-    }
-
-
-def split_into_batches(state: BatchState) -> BatchState:
-    """Resolve dependencies and split items into batches.
-
-    Args:
-        state: Current batch state
-
-    Returns:
-        Updated state with dependency levels
-    """
-    items = state["items"]
-    object_type = state["object_type"]
-
-    logger.info(f"Resolving dependencies for {len(items)} items")
-
-    # Get dependency summary
-    summary = get_dependency_summary(items, object_type)
-    logger.info(f"Dependency summary: {summary}")
-
-    # Sort into dependency levels
-    try:
-        dependency_levels = sort_items_by_dependencies(items, object_type)
-
-        logger.info(f"Split into {len(dependency_levels)} dependency levels:")
-        for i, level in enumerate(dependency_levels):
-            logger.info(f"  Level {i}: {len(level)} items")
-
-        return {
-            **state,
-            "dependency_levels": dependency_levels,
-            "current_batch_index": 0,
-        }
-
-    except ValueError as e:
-        logger.error(f"Dependency resolution failed: {e}")
-        return {
-            **state,
-            "dependency_levels": [],
-            "result_message": f"❌ Error: {e}",
-        }
-
-
-def check_and_process_next_batch(state: BatchState) -> list[SendType]:
-    """Check if there are more batches and fan out to parallel processing.
-
-    Uses LangGraph Send API to create parallel executions.
-
-    Args:
-        state: Current batch state
-
-    Returns:
-        List of Send commands for parallel execution
-    """
-    current_idx = state["current_batch_index"]
-    levels = state["dependency_levels"]
-
-    if current_idx >= len(levels):
-        # No more batches
-        return []
-
-    current_level = levels[current_idx]
-    logger.info(
-        f"Processing batch {current_idx + 1}/{len(levels)} "
-        f"with {len(current_level)} items"
-    )
-
-    # Create Send for each item in current level (parallel execution)
-    sends = []
-    for item in current_level:
-        # Prepare state for single item CRUD operation
-        item_state = {
-            "operation_type": state["operation_type"],
-            "object_type": state["object_type"],
-            "object_name": item.get("name"),
-            "data": item,
-            "validation_result": None,
-            "exists": None,
-            "operation_result": None,
-            "verification_result": None,
-            "message": "",
-            "error": None,
-        }
-
-        # Send to process_single_item node
-        sends.append(Send("process_single_item", item_state))
-
-    return sends
-
-
-def process_single_item(state: dict) -> dict:
-    """Process a single item using CRUD subgraph.
-
-    This node is invoked in parallel for each item in a batch level.
-
-    Args:
-        state: CRUD state for single item
-
-    Returns:
-        Updated state with execution result
-    """
-    item_name = state.get("object_name", "unknown")
-    logger.info(f"Processing item: {item_name}")
-
-    # Create CRUD subgraph
-    crud_graph = create_crud_subgraph()
-
-    try:
-        # Invoke CRUD subgraph
-        result = crud_graph.invoke(state, config={"configurable": {"thread_id": str(uuid.uuid4())}})
-
-        # Extract result for aggregation
-        return {
-            "name": item_name,
-            "status": "success" if "✅" in result.get("message", "") else "error",
-            "message": result.get("message", ""),
-            "error": result.get("error"),
-        }
-
-    except Exception as e:
-        logger.error(f"Error processing {item_name}: {e}")
-        return {
-            "name": item_name,
-            "status": "error",
-            "message": f"❌ Error: {e}",
-            "error": str(e),
-        }
-
-
-def aggregate_batch_results(state: BatchState) -> BatchState:
-    """Aggregate results from parallel batch execution.
-
-    Args:
-        state: Current batch state with current_batch_results
-
-    Returns:
-        Updated state with aggregated statistics
-    """
-    batch_results = state.get("current_batch_results", [])
-
-    if not batch_results:
-        logger.warning("No batch results to aggregate")
-        return state
-
-    logger.info(f"Aggregating {len(batch_results)} batch results")
-
-    # Count successes and failures
-    successful = sum(1 for r in batch_results if r.get("status") == "success")
-    failed = sum(1 for r in batch_results if r.get("status") == "error")
-
-    # Collect failure details
-    failures = [
-        {"name": r["name"], "error": r.get("error", "Unknown error")}
-        for r in batch_results
-        if r.get("status") == "error"
-    ]
-
-    return {
-        **state,
-        "completed_items": state["completed_items"] + len(batch_results),
-        "successful_items": state["successful_items"] + successful,
-        "failed_items": state["failed_items"] + failed,
-        "failure_details": state["failure_details"] + failures,
-        "current_batch_results": [],  # Clear for next batch
-        "current_batch_index": state["current_batch_index"] + 1,
-    }
-
-
-def should_continue(state: BatchState) -> Literal["check_and_process_next_batch", "format_batch_response"]:
-    """Determine if there are more batches to process.
-
-    Args:
-        state: Current batch state
-
-    Returns:
-        Next node name
-    """
-    current_idx = state["current_batch_index"]
-    total_levels = len(state.get("dependency_levels", []))
-
-    if current_idx < total_levels:
-        # More batches to process
-        logger.info(f"Continuing to batch {current_idx + 1}/{total_levels}")
-        return "check_and_process_next_batch"
-    else:
-        # All batches complete
-        logger.info("All batches complete")
-        return "format_batch_response"
-
-
-def format_batch_response(state: BatchState) -> BatchState:
-    """Format final batch operation response.
-
-    Args:
-        state: Current batch state
-
-    Returns:
-        Updated state with formatted message
-    """
-    total = state["total_items"]
-    completed = state["completed_items"]
-    successful = state["successful_items"]
-    failed = state["failed_items"]
-
-    # Build message
-    message_parts = [
-        f"📊 Batch {state['operation_type']} for {state['object_type']} - Summary",
-        f"",
-        f"Total items: {total}",
-        f"✅ Successful: {successful}",
-        f"❌ Failed: {failed}",
-        f"",
-    ]
-
-    if state.get("dependency_levels"):
-        levels = state["dependency_levels"]
-        message_parts.append(f"Dependency levels: {len(levels)}")
-        message_parts.append(
-            f"Items per level: {', '.join(str(len(l)) for l in levels)}"
-        )
-        message_parts.append("")
-
-    # Add failure details (first 5)
-    if state["failure_details"]:
-        message_parts.append("Failures:")
-        for failure in state["failure_details"][:5]:
-            message_parts.append(f"  - {failure['name']}: {failure['error']}")
-
-        if len(state["failure_details"]) > 5:
-            remaining = len(state["failure_details"]) - 5
-            message_parts.append(f"  ... and {remaining} more")
-
-    # Overall status
-    if failed == 0:
-        status_icon = "✅"
-        status_text = "All operations completed successfully"
-    elif successful > 0:
-        status_icon = "⚠️"
-        status_text = "Partial success - some operations failed"
-    else:
-        status_icon = "❌"
-        status_text = "All operations failed"
-
-    message_parts.insert(0, f"{status_icon} {status_text}")
-
-    message = "\n".join(message_parts)
-
-    return {**state, "result_message": message}
-
-
-def create_batch_subgraph() -> StateGraph:
-    """Create batch operations subgraph with parallel execution.
-
-    Returns:
-        Compiled StateGraph for batch operations
-    """
-    workflow = StateGraph(BatchState)
-
-    # Add nodes
-    workflow.add_node("validate_batch_input", validate_batch_input)
-    workflow.add_node("split_into_batches", split_into_batches)
-    workflow.add_node("check_and_process_next_batch", check_and_process_next_batch)
-    workflow.add_node("process_single_item", process_single_item)
-    workflow.add_node("aggregate_batch_results", aggregate_batch_results)
-    workflow.add_node("format_batch_response", format_batch_response)
-
-    # Add edges
-    workflow.set_entry_point("validate_batch_input")
-    workflow.add_edge("validate_batch_input", "split_into_batches")
-    workflow.add_edge("split_into_batches", "check_and_process_next_batch")
-
-    # Conditional: fan-out to parallel processing or finish
-    workflow.add_conditional_edges(
-        "check_and_process_next_batch",
-        lambda state: (
-            ["process_single_item"]
-            if state["current_batch_index"] < len(state.get("dependency_levels", []))
-            else []
-        ),
-        ["process_single_item"],
-    )
-
-    # After parallel processing, aggregate results
-    workflow.add_edge("process_single_item", "aggregate_batch_results")
-
-    # Conditional: continue to next batch or format response
-    workflow.add_conditional_edges(
-        "aggregate_batch_results",
-        should_continue,
-        {
-            "check_and_process_next_batch": "check_and_process_next_batch",
-            "format_batch_response": "format_batch_response",
-        },
-    )
-
-    # End after formatting
-    workflow.add_edge("format_batch_response", END)
-
-    return workflow.compile()
